refactor : #466 Payment 디렉터리 구조 변경
