+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/service/event/PaymentEventHandlerService.java
@@ -1,4 +1,4 @@
-package com.taken_seat.payment_service.application.service;
+package com.taken_seat.payment_service.application.service.event;
 
 import com.taken_seat.common_service.message.PaymentMessage;
